+++ src/hopkins_statistic/__init__.py
@@ -6,10 +6,15 @@ Functions:
 """
 
 __all__ = [
+    "Alternative",
+    "Frame",
     "HopkinsTestResult",
+    "ToRNG",
     "hopkins",
     "hopkins_test",
 ]
 
-from ._inference import HopkinsTestResult, hopkins_test
+from ._inference import Alternative, HopkinsTestResult, hopkins_test
+from ._sampling import Frame
 from ._statistic import hopkins
+from ._typing import ToRNG